Be explicit about introducing loops at ParZ level.
Previously, lifting to `ParZ` puts the lifted `Z` computation in an infinite
loop since this makes it easy to deal with shutting down related threads when
appropriate. However, we probably want to change this in the future, so we
should be explicit about this happening, requiring users to use `repeatz` to
lift their `Z` computations to the `ParZ` level.